+++ tests/PersistentModelTest.php
@@ -2,7 +2,6 @@
 
 namespace Rubix\ML\Tests;
 
-use Rubix\ML\AnomalyDetectors\Scoring;
 use Rubix\ML\Learner;
 use Rubix\ML\Wrapper;
 use Rubix\ML\DataType;
@@ -10,7 +9,9 @@ use Rubix\ML\Estimator;
 use Rubix\ML\Probabilistic;
 use Rubix\ML\EstimatorType;
 use Rubix\ML\PersistentModel;
+use Rubix\ML\Serializers\RBX;
 use Rubix\ML\Persisters\Filesystem;
+use Rubix\ML\AnomalyDetectors\Scoring;
 use Rubix\ML\Classifiers\DummyClassifier;
 use PHPUnit\Framework\TestCase;
 
@@ -30,7 +31,7 @@ class PersistentModelTest extends TestCase
      */
     protected function setUp() : void
     {
-        $this->estimator = new PersistentModel(new DummyClassifier(), new Filesystem('test.model'));
+        $this->estimator = new PersistentModel(new DummyClassifier(), new Filesystem('test.model'), new RBX());
     }
 
     /**
@@ -70,6 +71,7 @@ class PersistentModelTest extends TestCase
         $expected = [
             'base' => new DummyClassifier(),
             'persister' => new Filesystem('test.model'),
+            'serializer' => new RBX(),
         ];
 
         $this->assertEquals($expected, $this->estimator->params());